+++ src/App.js
@@ -1,10 +1,13 @@
 import logo from './logo.svg';
 import './App.css';
+import AllEvents from './components/allEvents'
 
 function App() {
   return (
     <div className="App">
-      
+      <AllEvents></AllEvents>
+
+
     </div>
   );
 }
