+++ src/redux/store.js
@@ -1,6 +1,9 @@
-import {createStore} from 'redux'
+import {createStore,applyMiddleware} from 'redux'
 import events from './reducer/allEvents'
+import {getEvents} from './middlweare/crud'
+
 const eventsReducer = events;
-const store = createStore(eventsReducer);
+const store = createStore(eventsReducer,applyMiddleware(getEvents));
 window.store = store;
+store.dispatch({ type: 'GET_DATA'});
 export default store;
